+++ fxapom/pages/base.py
@@ -2,18 +2,19 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
+from pypom import Page
+
 from fxapom.fxapom import TIMEOUT
-from .page import Page
 
 
 class Base(Page):
 
-    def __init__(self, driver, timeout=TIMEOUT):
-        super(Base, self).__init__(driver, timeout)
-        self._main_window_handle = self.driver.current_window_handle
+    def __init__(self, selenium, timeout=TIMEOUT):
+        super(Base, self).__init__(selenium, timeout=timeout)
+        self._main_window_handle = self.selenium.current_window_handle
 
     def switch_to_main_window(self):
-        self.driver.switch_to_window(self._main_window_handle)
+        self.selenium.switch_to_window(self._main_window_handle)
 
     def close_window(self):
-        self.driver.close()
+        self.selenium.close()