+++ fxapom/pages/sign_in.py
@@ -3,8 +3,7 @@
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 from selenium.webdriver.common.by import By
-from selenium.webdriver.support import expected_conditions as EC
-from selenium.webdriver.support.ui import WebDriverWait as Wait
+from selenium.webdriver.support import expected_conditions as expected
 
 from fxapom.fxapom import TIMEOUT
 from fxapom.pages.base import Base
@@ -18,37 +17,37 @@ class SignIn(Base):
     _password_input_locator = (By.ID, 'password')
     _sign_in_locator = (By.ID, 'submit-btn')
 
-    def __init__(self, driver, timeout=TIMEOUT):
-        super(SignIn, self).__init__(driver, timeout)
+    def __init__(self, selenium, timeout=TIMEOUT):
+        super(SignIn, self).__init__(selenium, timeout)
         self._sign_in_window_handle = None
         self.popup = False
-        self.check_for_popup(self.driver.window_handles)
+        self.check_for_popup(self.selenium.window_handles)
 
     @property
     def login_password(self):
         """Get the value of the login password field."""
-        return self.driver.find_element(*self._password_input_locator).get_attribute('value')
+        return self.selenium.find_element(*self._password_input_locator).get_attribute('value')
 
     @login_password.setter
     def login_password(self, value):
         """Set the value of the login password field."""
-        password = self.driver.find_element(*self._password_input_locator)
+        password = self.selenium.find_element(*self._password_input_locator)
         password.clear()
         password.send_keys(value)
 
     def click_next(self):
-        self.driver.find_element(*self._next_button_locator).click()
+        self.selenium.find_element(*self._next_button_locator).click()
 
     @property
     def email(self):
         """Get the value of the email field."""
-        return self.driver.find_element(*self._email_input_locator).get_attribute('value')
+        return self.selenium.find_element(*self._email_input_locator).get_attribute('value')
 
     @email.setter
     def email(self, value):
         """Set the value of the email field."""
-        email = Wait(self.driver, self.timeout).until(
-            EC.visibility_of_element_located(self._email_input_locator))
+        email = self.wait.until(expected.visibility_of_element_located(
+            self._email_input_locator))
         email.clear()
         email.send_keys(value)
 
@@ -56,19 +55,20 @@ class SignIn(Base):
         if len(handles) > 1:
             self.popup = True
             for handle in handles:
-                self.driver.switch_to.window(handle)
+                self.selenium.switch_to.window(handle)
                 if self.is_element_present(*self._fox_logo_locator):
                     self._sign_in_window_handle = handle
                     break
             else:
                 raise Exception('Popup has not loaded')
-        Wait(self.driver, self.timeout).until(EC.visibility_of_element_located(self._email_input_locator))
+        self.wait.until(expected.visibility_of_element_located(
+            self._email_input_locator))
 
     def click_sign_in(self):
-        self.driver.find_element(*self._sign_in_locator).click()
+        self.selenium.find_element(*self._sign_in_locator).click()
         if self.popup:
-            Wait(self.driver, self.timeout).until(
-                lambda s: self._sign_in_window_handle not in self.driver.window_handles)
+            self.wait.until(
+                lambda s: self._sign_in_window_handle not in self.selenium.window_handles)
             self.switch_to_main_window()
 
     def sign_in(self, email, password):
@@ -76,7 +76,7 @@ class SignIn(Base):
         self.email = email
         self.login_password = password
         if self.is_element_present(*self._next_button_locator):
-            Wait(self.driver, self.timeout).until(
-                EC.visibility_of_element_located(self._next_button_locator))
+            self.wait.until(expected.visibility_of_element_located(
+                self._next_button_locator))
             self.click_next()
         self.click_sign_in()